Initial version of fissile development playground
The repository contains the simplest possible configuration to run fissile
commands. It can be used to test changes done to fissile.

This is intended to show case the typical fissile flow from creating a BOSH
release to generating a set of Helm Charts. The actual result may be deployable
to a Kubernetes cluster, but it was never tested to run or produce any
meaningful results.

Signed-off-by: EEE <encalada@de.ibm.com>